Refactoring method names in DbStore classes (#90)
The calling semantics of the DbStore classes are a bit redundant because we
end up with calls like 'temperature_store.store_temperature'. The class name
gives the type, so we don't have to repeat it in the method. I think if we do
this, it opens up the door to some refactoring around managing multiple
db store instances of different types.