Changing timestamp serialization for database (#101)
dateutil's parsing is apparently very slow because it has to do a lot of
regexing on the date string to figure out its format automatically. We don't
strictly need this because we can just serialize it to a format that's
easier to parse back.

This changes the date serialization so that dates are always stored as UTC
times (so that we don't have to try to serialize/deserialize time zone). This
also removes milliseconds precision from the serialized format because it
wasn't a deliberate choice to add it and the timestamps probably aren't
precise to the millisecond of when the sensor was read.

This should fix #98.