Fixing measurement bug in soil moisture sensor (#125)
The second reading was coming in wrong because we weren't turning off both
GPIO pins before we slept, so there was probably some remnant charge on the
first pin during the second read. Adding in a turn_pin_off call to ensure
that both pins are off and have time for the charge to dissipate.