Fixing last_poll_time function (#137)
* Fixing last_poll_time function

last_poll_time was returning timestamps as ints, whereas the callers were
expecting datetimes. Also, the way that last_poll_time was getting updated
wasn't really intuitive, so updated the code so that last_poll_time only
updates after the scheduler waits through a poll boundary.

* Making a clearer distinction in Scheduler between UNIX time and datetime

The Scheduler uses both UNIX timestamps and datetimes, but it's not always
clear which variable/function is using which format. This can lead to bugs
where types are used or compared incorrectly. Changing the naming so that UNIX
timestamps are always named to distinguish from normal datetimes.